Center pointer for IconBelowCursor placement
When opening the menu with the keyboard, it's helpful to
have the mouse pointer in the middle of the screen so the
pointer is always exactly above the focused icon.

Otherwise, when the menu is near the edges of the screen,
the menu is shifted to make it fully visible what means
the mouse pointer is not exactly above the focused icon.

Another way to solve this problem would be to allow the
menu to expand beyond the screen edges when the menu is
opened with the keyboard.

Also, there should probably be a settings flag for this.